cleanup/fix gui code rename symbols for consistency add new macros for conversion between event code and constrol id to do code more readable Use control id as much as possible and calculate event code from control id if necessary add macros for MDI windows and menu controls

+++ bld/gui/win/c/guitool.c
@@ -352,7 +352,7 @@ bool GUIXCreateToolBarWithTips( gui_window *wnd, bool fixed, gui_ord height,
 
     for( i = 0; i < num_toolbar_items; i++ ) {
         info.u.bmp = toolbar->bitmaps[i];
-        info.id = (WORD)toolinfo[i].id;
+        info.id = toolinfo[i].id;
         info.flags = 0;
         if( use_tips && toolinfo[i].tip != NULL ) {
             strncpy( info.tip, toolinfo[i].tip, MAX_TIP );
